refactor
* moved compare_json to process module
* expose process module as public
